[hal] common filesystem/inflate implementation
